+++ urbanairship-core/src/main/java/com/urbanairship/AirshipConfigOptions.java
@@ -413,6 +413,12 @@ public class AirshipConfigOptions {
      */
     public final boolean isPromptForPermissionOnUserNotificationsEnabled;
 
+    /**
+     * Flag indicating whether or not the SDK will automatically pause In-App Automation during app launch.
+     * Defaults to false.
+     */
+    public final boolean autoPauseInAppAutomationOnLaunch;
+
     private AirshipConfigOptions(@NonNull Builder builder) {
         if (builder.inProduction) {
             this.appKey = firstOrEmpty(builder.productionAppKey, builder.appKey);
@@ -466,6 +472,7 @@ public class AirshipConfigOptions {
         this.fcmFirebaseAppName = builder.fcmFirebaseAppName;
         this.initialConfigUrl = builder.initialConfigUrl;
         this.isPromptForPermissionOnUserNotificationsEnabled = builder.isPromptForPermissionOnUserNotificationsEnabled;
+        this.autoPauseInAppAutomationOnLaunch = builder.autoPauseInAppAutomationOnLaunch;
     }
 
     private static <T> List<T> copyOrEmpty(@Nullable List<T> list) {
@@ -632,8 +639,8 @@ public class AirshipConfigOptions {
         private static final String FIELD_REQUIRE_INITIAL_REMOTE_CONFIG_ENABLED = "requireInitialRemoteConfigEnabled";
         private static final String FIELD_ENABLED_FEATURES = "enabledFeatures";
         private static final String FIELD_INITIAL_CONFIG_URL = "initialConfigUrl";
-
         private static final String FIELD_IS_PROMPT_FOR_PERMISSION_ON_USER_NOTIFICATIONS_ENABLED = "isPromptForPermissionOnUserNotificationsEnabled";
+        private static final String FIELD_AUTO_PAUSE_IN_APP_AUTOMATION_ON_LAUNCH = "autoPauseInAppAutomationOnLaunch";
 
         private String appKey;
         private String appSecret;
@@ -680,6 +687,8 @@ public class AirshipConfigOptions {
         private String initialConfigUrl;
 
         private boolean isPromptForPermissionOnUserNotificationsEnabled = true;
+
+        private boolean autoPauseInAppAutomationOnLaunch = false;
 
         /**
          * Apply the options from the default properties file {@code airshipconfig.properties}.
@@ -1048,6 +1057,10 @@ public class AirshipConfigOptions {
 
                         case FIELD_IS_PROMPT_FOR_PERMISSION_ON_USER_NOTIFICATIONS_ENABLED:
                             this.setIsPromptForPermissionOnUserNotificationsEnabled(configParser.getBoolean(name, true));
+                            break;
+
+                        case FIELD_AUTO_PAUSE_IN_APP_AUTOMATION_ON_LAUNCH:
+                            this.setAutoPauseInAppAutomationOnLaunch(configParser.getBoolean(name, false));
                             break;
 
                         case FIELD_ENABLED_FEATURES:
@@ -1648,6 +1661,19 @@ public class AirshipConfigOptions {
         @NonNull
         public Builder setIsPromptForPermissionOnUserNotificationsEnabled(boolean enabled) {
             this.isPromptForPermissionOnUserNotificationsEnabled = enabled;
+            return this;
+        }
+
+
+        /**
+         * Set the auto pause In-App Automation on launch.
+         *
+         * @param autoPauseInAppAutomationOnLaunch {@code true} to auto pause In-App Automation, otherwise {@code false}.
+         * @return The config options builder.
+         */
+        @NonNull
+        public Builder setAutoPauseInAppAutomationOnLaunch(boolean autoPauseInAppAutomationOnLaunch) {
+            this.autoPauseInAppAutomationOnLaunch = autoPauseInAppAutomationOnLaunch;
             return this;
         }