+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 [Migration Guides](https://github.com/urbanairship/android-library/tree/main/documentation/migration)
 
-## Version 17.0.0, May 26, 2023
+## Version 17.0.0-Beta, May 26, 2023
 
 Beta release for SDK 17.0.0. This release brings several breaking changes. This version is not suitable for a production app, but we encourage testing out the new APIs and providing us feedback so we can make changes before the final SDK 17 release.
 
@@ -14,6 +14,12 @@ Beta release for SDK 17.0.0. This release brings several breaking changes. This 
 - Log listener has been removed and replaced by a log handler.
 - Channel listener has been updated to only listen for channel create, channel update has been removed
 - Added new PushNotificationStatus and PushNotificationStatus flow that provides the current status of push notifications
+
+## Version 16.11.0, June 12, 2023
+Minor release that adds new config option `autoPauseInAppAutomationOnLaunch` to always pause IAA during app launch. 
+
+### Changes
+- Added `autoPauseInAppAutomationOnLaunch` config option
 
 ## Version 16.10.0, June 2, 2023